[IMP] stock_picking_batch_creation_group_reservation_rate: Use abstract model
