+++ stock_picking_batch_creation_group_reservation_rate/__init__.py
@@ -1,1 +1,1 @@
-from . import wizards
+from . import models, wizards
