+++ stock_picking_batch_creation_group_reservation_rate/models/__init__.py
@@ -1,0 +1,1 @@
+from . import picking_batch_abstract
